requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
